Moved install part of npm into a child-process.
This stops npm being so shouty when we're installing deps
