+++ lib/npm.js
@@ -3,7 +3,6 @@
 var cp = require('child_process'),
     Q = require('q'),
     npm = require('npm'),
-    fs = require('fs'),
     path = require('path'),
     packagePath = path.resolve(process.cwd(), './package.json');
 
@@ -31,8 +30,7 @@ module.exports.list = function (packageJson, devFlag) {
 
 // make npm quiet for outdated command using a child process
 module.exports.outdated = function (packages) {
-    var deferred = Q.defer(),
-        timeout;
+    var deferred = Q.defer();
 
     //to shush npm logs
     var outdated = cp.fork(__dirname + '/outdated', {
@@ -63,40 +61,25 @@ module.exports.install = function (packagesToUpdate, options) {
     if (typeof packagesToUpdate === 'string') {
         packagesToUpdate = [packagesToUpdate];
     }
+
     var deferred = Q.defer();
 
-    npm.load('../package', function () {
-        if (options.save) {
-            if (options.dev) {
-                npm.config.set('save-dev', true);
-            } else {
-                npm.config.set('save', true);
-            }
-        }
-        npm.commands.install(packagesToUpdate, function (err) {
-            if (!err) {
-                if (options.save) {
-                    try {
-                        var packageJSON = JSON.parse(fs.readFileSync(packagePath, 'utf8'));
-                        var dependencyType = options.dev ? 'devDependencies' : 'dependencies';
-                        packagesToUpdate.forEach(function (pack) {
-                            var packParts = pack.split('@');
-                            var packName = packParts[0];
-                            var semVerRange = packParts[1].match(/[~^]/g) || '';
-                            packageJSON[dependencyType][packName] = packageJSON[dependencyType][packName].replace(/[~^]/g, '');
-                            packageJSON[dependencyType][packName] = semVerRange + packageJSON[dependencyType][packName];
-                        });
-                        fs.writeFileSync(packagePath, JSON.stringify(packageJSON, null, 2));
-                    } catch (err) {
-                        deferred.reject(err);
-                    }
-                }
-                deferred.resolve();
-            } else {
-                deferred.reject(err);
-            }
-        });
+    //to shush npm logs
+    var install = cp.fork(__dirname + '/install', {
+        cwd: process.cwd(),
+        silent: true
     });
+
+    install.on('message', function () {
+        deferred.resolve();
+    });
+
+    install.on('error', function (err) {
+        deferred.reject(err);
+    });
+
+
+    install.send({packagesToInstall: packagesToUpdate, options: options});
 
     return deferred.promise;
 };
