+++ test/lib/npm.test.js
@@ -4,49 +4,6 @@ var npmWrapper = require('../../lib/npm'),
     npm = require('npm');
 
 describe('npm()', function () {
-    describe('npm.install()', sandbox(function () {
-        var outdatedPackages = ['npm@0.0.3', 'q@0.0.4'],
-            deferred, stub;
-
-        beforeEach(function () {
-            this.stub(npm, 'load').yields();
-            npm.commands = {
-                install: this.stub().yields()
-            };
-            npm.config = {
-                set: this.stub()
-            };
-
-            stub = this.stub();
-            deferred = npmWrapper.install(outdatedPackages, {save: true}).then(stub);
-        });
-
-        afterEach(function () {
-            npm.load.restore();
-        });
-
-        it('calls npm.load()', function () {
-            npm.load.should.have.been.calledOnce;
-        });
-
-        describe('npm.load()', function () {
-            it('should call config.set() with save if save flag is passed', function () {
-                npm.config.set.should.have.been.calledWith('save', true);
-            });
-
-            it('should call npm.install()', function () {
-                npm.commands.install.should.have.been.calledOnce;
-                npm.commands.install.should.have.been.calledWith(outdatedPackages);
-            });
-
-            describe('npm.install()', function () {
-                it('should resolve the promise on success', function () {
-                    stub.should.have.been.calledOnce;
-                });
-            });
-        });
-    }));
-
     describe('npm.test()', sandbox(function () {
         var deferred, stub;
